+++ config/filenames.py
@@ -1,0 +1,8 @@
+import os.path
+
+PROGRAM_DATA_DIR = "program_data"
+
+CARD_PRISONER_SAVE_FILE = os.path.join(
+    PROGRAM_DATA_DIR,
+    "card_prisoner.dta",
+)