+++ config/key_bindings.py
@@ -8,6 +8,10 @@ confirm = K_z
 cancel = K_x
 cheat = K_c
 restart = K_r
+save = K_s
+load = K_l
+quit = K_q
+show_help = K_h
 
 # game specific: card prisoner
 draw_card = K_d